Move metadata structs to dedicated header
